gpg: properly locate key fingerprint
The gnupg list_keys() method was called with the key fingerprint as a
parameter, but this unfortunately just evaluates to True in the python
gnupg code which just lists all keys that have private keys available.

Properly fix by iterating through all public keys and locate the proper
key fingerprint.