gpg: some uids are latin1, some utf-8
Some key uids are latin1 and some are utf-8. To avoid decoding
exceptions, we tell python-gnupg to assume latin1 encoding.
Unfortunately, this also means some uids will not display correctly.

To work around it, try to convert from latin1 to utf-8 and if it fails
assume it was utf-8 to begin with. Hopefully, this should work in most
cases.