pwmanager: add pipe command
pwmanager pipe will print only the password for the unique host/user
match provided on the command line. If there are no matches or if the
host/user combination matches multiple entries, an error is printed on
stderr and pwmanager will exit with a non-zero exit status.

This is useful for piping into e.g. xclip.