ldap: exit with error if ca_cert file is not readable
Currently, pwmanager exits with a backtrace if the file containing the
CA certificate (as configured in pwmanager.conf) is not readable. Since
one could use a configuration example and the path to the CA
certificates vary between distributions, it makes sense to print a more
readable error message if this happens.