Provide gpg_path config directive
On some platforms, the gpg binary can be configured to launch either
gpg1 or gpg2. The python3-gnupg only works with gpg2, and thus sometimes
this binary is called 'gpg' and sometimes 'gpg2'.

Provide a configuration directive so this can be changed on a
per-installation basis. If omitted, the default 'gpg' without an
absolute path is used.