pwmanager: fix crash when key has been removed
If a key is removed and cannot be imported into the temporary keyring
when encrypting a file (such as when synchronizing the passwords),
pwmanager crashed stating the key was not found.

Fix this by keeping track of which keys exist and which don't, and
always reencrypt if a key has gone missing.